Implementação da validação dos formulários de plano e cadastro de clientes

+++ src/main/java/br/com/ricardotulio/mikrotikadmin/controller/PlanosController.java
@@ -12,26 +12,33 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import br.com.ricardotulio.mikrotikadmin.dao.PlanoDao;
+import br.com.ricardotulio.mikrotikadmin.dao.RadGroupReplyDao;
 import br.com.ricardotulio.mikrotikadmin.model.Plano;
+import br.com.ricardotulio.mikrotikadmin.model.RadGroupReply;
 
 @Controller
 public class PlanosController {
 
 	private PlanoDao planoDao;
-	
+
+	private RadGroupReplyDao radGroupReplyDao;
+
 	private static final String PLANO_CADASTRADO_COM_SUCESSO = "Plano cadastrado encontrado!";
 	private static final String PLANO_NAO_ENCONTRADO = "Plano não encontrado!";
 	private static final String PLANO_ATUALIZADO_COM_SUCESSO = "Plano atualizado encontrado!";
 	private static final String PLANO_EXCLUIDO_COM_SUCESSO = "Plano excluído encontrado!";
+	private static final String PLANO_NAO_EXCLUIDO_POSSUI_CLIENTES = "Plano não pode ser excluído pois possui clientes vinculados.";
 
 	@Autowired
-	public PlanosController(PlanoDao planoDao) {
+	public PlanosController(PlanoDao planoDao, RadGroupReplyDao radGroupReplyDao) {
 		this.planoDao = planoDao;
+		this.radGroupReplyDao = radGroupReplyDao;
 	}
-	
+
 	@RequestMapping(value = "/planos/", method = RequestMethod.GET)
 	public String index(Model model) {
 		List<Plano> planos = this.planoDao.obtemLista();
+				
 		model.addAttribute("planos", planos);
 		return "planos/index";
 	}
@@ -45,6 +52,13 @@ public class PlanosController {
 	@Transactional
 	public String cadastrarPost(Plano plano, final RedirectAttributes redirectAttributes) {
 		this.planoDao.persiste(plano);
+		
+		RadGroupReply radGroupReply = new RadGroupReply();
+		radGroupReply.setId(plano.getId());
+		radGroupReply.setGroupname(plano.getTitulo().replaceAll("\\s", "").toLowerCase());
+		radGroupReply.setValue(Integer.toString((int) (plano.getTaxaUpload() * 1024)) + "k/" + Integer.toString((int) (plano.getTaxaDownload() * 1024)) + "k");
+		this.radGroupReplyDao.persist(radGroupReply);
+		
 		redirectAttributes.addFlashAttribute("success", PlanosController.PLANO_CADASTRADO_COM_SUCESSO);
 		return "redirect:/planos/";
 	}
@@ -52,8 +66,8 @@ public class PlanosController {
 	@RequestMapping(value = "/planos/editar/{id}", method = RequestMethod.GET)
 	public String editarGet(@PathVariable(value = "id") Long id, Model model,
 			final RedirectAttributes redirectAttributes) {
-		Plano plano = this.planoDao.obtem(id);	
-		
+		Plano plano = this.planoDao.obtem(id);
+
 		if (plano == null) {
 			return this.informaQuePlanoNaoExiste(redirectAttributes);
 		}
@@ -64,13 +78,18 @@ public class PlanosController {
 
 	@RequestMapping(value = "/planos/editar/{id}", method = RequestMethod.POST)
 	@Transactional
-	public String editarPost(@PathVariable(value = "id") Long id, Plano plano, final RedirectAttributes redirectAttributes) {
-		if(this.planoDao.obtem(id) == null) {
+	public String editarPost(@PathVariable(value = "id") Long id, Plano plano,
+			final RedirectAttributes redirectAttributes) {
+		if (this.planoDao.obtem(id) == null) {
 			return this.informaQuePlanoNaoExiste(redirectAttributes);
 		}
-		
+
 		plano.setId(id);
 		this.planoDao.persiste(plano);
+		
+		RadGroupReply radGroupReply = radGroupReplyDao.get(plano.getId());
+		radGroupReply.setValue(Integer.toString((int) (plano.getTaxaUpload() * 1024)) + "k/" + Integer.toString((int) (plano.getTaxaDownload() * 1024)) + "k");
+		
 		redirectAttributes.addFlashAttribute("success", PlanosController.PLANO_ATUALIZADO_COM_SUCESSO);
 		return "redirect:/planos/";
 	}
@@ -79,12 +98,21 @@ public class PlanosController {
 	@Transactional
 	public String excluirGet(@PathVariable("id") Long id, final RedirectAttributes redirectAttributes) {
 		Plano plano = this.planoDao.obtem(id);
-		
-		if(plano == null) {
+
+		if (plano == null) {
 			return this.informaQuePlanoNaoExiste(redirectAttributes);
 		}
 		
+		if(this.planoDao.planoPossuiClientes(plano)) {
+			redirectAttributes.addFlashAttribute("error", PlanosController.PLANO_NAO_EXCLUIDO_POSSUI_CLIENTES);
+			return "redirect:/planos/";
+		}
+
 		this.planoDao.remove(plano);
+		
+		RadGroupReply radGroupReply = radGroupReplyDao.get(plano.getId());
+		radGroupReplyDao.remove(radGroupReply);
+		
 		redirectAttributes.addFlashAttribute("success", PlanosController.PLANO_EXCLUIDO_COM_SUCESSO);
 		return "redirect:/planos/";
 	}
@@ -94,3 +122,4 @@ public class PlanosController {
 		return "redirect:/planos/";
 	}
 }
+
